Moved from two passes to almost-single-pass
It makes a full first pass, collecting labels, and lines with
unresolved references. These are assembled anyway in order to
determine the resuting code size, but the result is overwritten later
in the second pass, in which the lines with undefined references are
recompiled in place.

If during the recompilation pass, references are still undefined, a
warning will be issued.